day15: make part1 much faster
Before, we were keeping a list of all unexplored points, even though
most of them were just infinite, and so would never be the next point to
explore.

Modify this to just add to that list whenever we first explore it.

Ideally we'd use a priority queue, but I'm not coding that up unless I
actually need to.