day15: switch to tree-based heap
Which, in practice, is faster. The reason, I believe, is because nesting
attrs as a tree lets us do partial updates of the heap (i.e. update only
the lhs w/o changing the rhs reference), while any change to a list
results in a full copy of the entire list.

Basically, immutable lists are iffy.